+++ src/p_local.h
@@ -269,6 +269,7 @@ extern mobj_t       **blocklinks;   // for thing chains
 #define MAXHEALTH   100
 
 void P_TouchSpecialThing(mobj_t *special, mobj_t *toucher, dboolean message, dboolean stat);
+dboolean P_TakeSpecialThing(mobjtype_t type);
 
 void P_DamageMobj(mobj_t *target, mobj_t *inflicter, mobj_t *source, int damage, dboolean adjust);